"Move CH13-Exceptions-Demo project to project root"
The files for CH13-Exceptions-Demo project were previously placed in a subdirectory. The commit here represents moving all file contents related to the project from the CH13-Exceptions-Demo subdirectory to the project root.

This restructuring was done to clean up the project structure and make the navigation easier for developers. Now, all relevant files are directly accessible from the root directory.