"Add zero and negative input handling to MilesPerGallon class"
In this commit, we have introduced input validations for the 'MilesPerGallon' class. Two new methods 'ValidateNonNegative' and 'ValidateNonZero' have been added to validate the miles and gallons input values. If the conditions are not met, appropriate exceptions are thrown. This ensures that the MilesPerGallon object is always in a valid state. We have also created a set of NUnit tests to verify these newly introduced checks. Further, a new exception class 'CannotBeZeroException' is added to handle cases where the provided value cannot be zero.